Finished working on activation and confirmation

+++ src/app/Components/start/recovery-password-change/recovery-password-change.component.ts
@@ -19,6 +19,8 @@ export class RecoveryPasswordChangeComponent implements OnInit {
 
   noInternet: Boolean = false;
   pageLoading: Boolean = true;
+
+  error: String = "";
   //pageLoading: Boolean = false;
 
   constructor(private tokenService: TokenService, private router: Router, private accountService: AccountService) {
@@ -54,6 +56,7 @@ export class RecoveryPasswordChangeComponent implements OnInit {
     this.requestError = false;
     this.progressbar = true;
     this.passwordChanged = false;
+    this.error = "";
 
     if(this.tokenService.doResetTokenExist()){
       var resetJson = this.tokenService.getResetTokenObject();
@@ -61,24 +64,57 @@ export class RecoveryPasswordChangeComponent implements OnInit {
         var RESET_TOKEN = resetJson["RESET_TOKEN"];
         var RESET_EMAIL = resetJson["RESET_EMAIL"];
 
-
+        //RESET_TOKEN = RESET_TOKEN + "1231";
       this.accountService.reset_change_password(RESET_TOKEN, RESET_EMAIL, newPassword)
       .subscribe(
         data => {
-          alert(data["RETURN_CODE"]);
-          this.passwordChanged = true;
+          //alert(data["RETURN_CODE"]);
+
+          var RETURN_CODE = data["RETURN_CODE"];
+          if(RETURN_CODE == 1){
+            // success
+            this.requestError = false;
+            this.passwordChanged = true;
+            this.tokenService.removeResetToken();
+          }
+          else if(RETURN_CODE == 2){
+            // exception in controller
+            this.requestError = true;
+            this.error = "Our servers encountered some internal error";
+          }
+          else if(RETURN_CODE == 3){
+            // reset token is not valid
+            this.requestError = true;
+            this.error = "Please refresh the page";
+          }
+          else if(RETURN_CODE == 4){
+            // password not changed
+            this.requestError = true;
+            this.error = "Unable to change password right now, try later.";
+          }
+          else if(RETURN_CODE == 5){
+            // New password is same as old password
+            this.requestError = true;
+            this.error = "You have entered your current password";
+          }
+          else{
+            this.requestError = true;
+            this.error = "Some unknown error has occured";
+          }
+
+          
           this.progressbar = false;
           
-          this.tokenService.removeResetToken();
 
         },error => {
           alert("ERROR");
+          this.error = "Unable to connect our services";
           this.requestError = true;
           this.progressbar = false;
         },
         () => {
           this.progressbar = false;
-          this.requestError = false;
+          //this.requestError = false;
         }
       );
